resolve an issue where react router does not like Routes wrapped in design system provider
