+++ packages/fast-component-explorer/app/app.tsx
@@ -1,21 +1,9 @@
 import React from "react";
-import Explorer from "./explorer";
+import ExplorerRegion from "./explorer-region";
 import Preview from "./preview";
-import { DesignSystemProvider } from "@microsoft/fast-jss-manager-react";
-import {
-    DesignSystem,
-    DesignSystemDefaults,
-} from "@microsoft/fast-components-styles-msft";
 import { history, initialComponentRoute } from "./config";
 import { Redirect, Route, Router, Switch } from "react-router-dom";
 
-const creatorDesignSystem: DesignSystem = Object.assign({}, DesignSystemDefaults, {
-    density: -2,
-    backgroundColor:
-        DesignSystemDefaults.neutralPalette[
-            DesignSystemDefaults.neutralPalette.length - 1
-        ],
-});
 /**
  * The root level app
  *
@@ -26,13 +14,11 @@ export default class App extends React.Component<{}, {}> {
         return (
             <Router history={history}>
                 <Switch>
-                    <DesignSystemProvider designSystem={creatorDesignSystem}>
-                        <Route
-                            component={Explorer}
-                            exact={true}
-                            path="/components/:component"
-                        />
-                    </DesignSystemProvider>
+                    <Route
+                        component={ExplorerRegion}
+                        exact={true}
+                        path="/components/:component"
+                    />
                     <Route component={Preview} exact={true} path="/preview" />
                     <Route>
                         <Redirect to={initialComponentRoute} />
